Moved background tasks to the tasks folder
